requirements.txt
fastapi==0.115.4
pre-commit==4.0.1
pydantic-settings==2.6.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2==2.9.10
alembic==1.13.3
asyncpg==0.30.0
fastapi-users[sqlalchemy,oauth]
python-jose==3.3.0
passlib==1.7.4